Issue edina/fieldtrip-authoring-tool/issues/30: Initial code for populating popups during animation.
NOTE: This only works if you first click on the Memories viewer and load the tracks.  It used to work
just going straight to the track animator, but this behaviour seems to have changed in commit 08d5268

Also, note that it takes a while for images to load (and they load rotated 90 degrees for some reason).  This
might not be so apparent if we switch from using dropbox to a local store.  It may be we should cache all this on
page load in any case.